Introduced traccc::device::make_doublet_buffers.
Creating the doublet buffers from the counts stored in
traccc::device::doublet_counter_buffer is not a completely
trivial operation. This function helps with doing it correctly.

Due to the setup of the seed finding algorithms, they currently
repeat some operations themselves. But those will be possible
to remove once the rest of the kernels will be updated.